requirements.txt
aioredis==1.2.0
appnope==0.1.0
asgiref==2.3.2
astroid==2.2.0
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
autobahn==19.2.1
Automat==0.7.0
autopep8==1.4.3
backcall==0.1.0
certifi==2018.11.29
channels==2.1.7
channels-redis==2.3.3
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.2
daphne==2.2.5
decorator==4.3.2
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.7
django-allauth==0.38.0
django-extra-fields==0.3
django-filter==2.1.0
django-heroku==0.3.1
django-rest-auth==0.9.3
django-rest-swagger==2.0.7
djangorestframework==3.9.1
djangorestframework-gis==0.14
factory-boy==2.11.1
Faker==1.0.2
hiredis==1.0.0
httpie==1.0.2
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
ipdb==0.11
ipython==7.3.0
ipython-genutils==0.2.0
isort==4.3.9
itypes==1.1.0
jedi==0.13.3
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
mockredispy==2.9.3
more-itertools==6.0.0
msgpack==0.6.1
oauthlib==3.0.1
onesignal-sdk==0.1.1
openapi-codec==1.3.2
parso==0.3.4
pbr==5.1.1
pep8==1.7.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pluggy==0.9.0
prompt-toolkit==2.0.9
psycopg2==2.7.7
psycopg2-binary==2.7.7
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
Pygments==2.3.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.3.0
pylint-django==2.0.6
pylint-plugin-utils==0.5
pytest==4.3.1
pytest-asyncio==0.10.0
pytest-django==3.4.8
python-dateutil==2.8.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.9
requests==2.21.0
requests-oauthlib==1.2.0
selenium==3.141.0
simplejson==3.16.0
six==1.12.0
social-auth-core==3.0.0
stevedore==1.30.0
text-unidecode==1.2
traitlets==4.3.2
Twisted==18.9.0
txaio==18.8.1
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.1
virtualenv==16.2.0
virtualenv-clone==0.5.0
virtualenvwrapper==4.8.2
wcwidth==0.1.7
whitenoise==4.1.2
wrapt==1.11.1
zope.interface==4.6.0
